Remove Docker - use local PHP for development instead
Reverts 6bf02cf681db48003701f317cc52d73e2c4bc47a and several other commits

I thought using Docker would make it easier to run multiple versions of PHP.
In reality it just made development slower and harder.
Having to write a script for every app was tedious.
Having to wait while every dependency installed was also.
So it's back to Plan A.

For my current development environment see:
https://github.com/davejamesmiller/dev-ansible
(Basically Ubuntu 18.04 + Apache 2.4 + PHP 7.2 & 7.1 FPM)

Unfortunately that also means putting the test app in a separate repo, because
Composer doesn't like it when the test project is in a directory below the package.
https://github.com/composer/composer/issues/6085
(In Docker I worked around this with a second mount point.)